Move ticket status generator to utils
