+++ oscar_support/defaults.py
@@ -1,4 +1,3 @@
-from django.db.models import get_model
 from django.utils.translation import ugettext_lazy as _
 
 SUPPORT_INITIAL_STATUS = _("New")
@@ -6,9 +5,6 @@ SUPPORT_INITIAL_STATUS_SLUG = "new"
 
 SUPPORT_RESOLVED_STATUS = _("Resolved")
 SUPPORT_RESOLVED_STATUS_SLUG = "resolved"
-
-TicketStatus = get_model("oscar_support", "TicketStatus")
-
 
 # Setting defaults for the REST API provided by django-rest-framework
 # for security reasons we only enable session-based authentication and
@@ -25,29 +21,3 @@ REST_FRAMEWORK = {
         'rest_framework.renderers.JSONRenderer',
     )
 }
-
-
-def get_ticket_initial_status():
-    try:
-        initial_status = TicketStatus.objects.get(
-            slug=SUPPORT_INITIAL_STATUS_SLUG
-        )
-    except TicketStatus.DoesNotExist:
-        initial_status = TicketStatus.objects.create(
-            slug=SUPPORT_INITIAL_STATUS_SLUG,
-            name=SUPPORT_INITIAL_STATUS
-        )
-    return initial_status
-
-
-def get_ticket_resolved_status():
-    try:
-        resolved_status = TicketStatus.objects.get(
-            slug=SUPPORT_RESOLVED_STATUS_SLUG
-        )
-    except TicketStatus.DoesNotExist:
-        resolved_status = TicketStatus.objects.create(
-            slug=SUPPORT_RESOLVED_STATUS_SLUG,
-            name=SUPPORT_RESOLVED_STATUS
-        )
-    return resolved_status
