+++ oscar_support/forms/__init__.py
@@ -2,7 +2,7 @@ from django import forms
 from django.db.models import get_model
 from django.utils.translation import ugettext_lazy as _
 
-from .. import defaults
+from .. import utils
 
 Order = get_model('order', 'Order')
 RelatedOrder = get_model("oscar_support", "RelatedOrder")
@@ -26,7 +26,7 @@ class TicketCreateForm(forms.ModelForm):
     def save(self, commit=True):
         instance = super(TicketCreateForm, self).save(commit=False)
 
-        instance.status = defaults.get_ticket_initial_status()
+        instance.status = utils.TicketStatusGenerator.get_initial_status()
         instance.requester = self.user
 
         if commit:
