+++ oscar_support/views.py
@@ -1,27 +1,14 @@
 from django.views import generic
-from django.conf import settings
 from django.db.models import get_model
 from django.http import HttpResponseRedirect
 from django.core.urlresolvers import reverse
 
-from . import defaults
+from . import utils
 from .forms import TicketUpdateForm, TicketCreateForm
 
 Ticket = get_model('oscar_support', 'Ticket')
 Message = get_model('oscar_support', 'Message')
 TicketStatus = get_model('oscar_support', 'TicketStatus')
-
-SUPPORT_RESOLVED_STATUS = getattr(
-    settings,
-    "SUPPORT_RESOLVED_STATUS",
-    defaults.SUPPORT_RESOLVED_STATUS
-)
-
-SUPPORT_RESOLVED_STATUS_SLUG = getattr(
-    settings,
-    "SUPPORT_RESOLVED_STATUS_SLUG",
-    defaults.SUPPORT_RESOLVED_STATUS_SLUG
-)
 
 
 class TicketListView(generic.ListView):
@@ -38,7 +25,7 @@ class TicketListView(generic.ListView):
 
     def get_context_data(self, **kwargs):
         ctx = super(TicketListView, self).get_context_data(**kwargs)
-        resolved = defaults.get_ticket_resolved_status()
+        resolved = utils.TicketStatusGenerator.get_resolved_status()
         ctx['open_ticket_list'] = self.get_queryset().exclude(status=resolved)
         ctx['resolved_ticket_list'] = self.get_queryset().filter(
             status=resolved
