+++ tests/unit/test_utils.py
@@ -2,8 +2,9 @@ import mock
 
 from django.test import TestCase
 
-from oscar_support.models import Ticket
-from oscar_support.utils import TicketNumberGenerator
+from oscar_support import defaults
+from oscar_support.models import Ticket, TicketStatus
+from oscar_support.utils import TicketNumberGenerator, TicketStatusGenerator
 
 
 class TestTheTicketNumberGenerator(TestCase):
@@ -26,3 +27,30 @@ class TestTheTicketNumberGenerator(TestCase):
         number_kwargs = TicketNumberGenerator.generate_subticket_id(ticket)
         self.assertEquals(number_kwargs['number'], ticket.number)
         self.assertEquals(number_kwargs['subticket_id'], 1)
+
+
+class TestTheStatusGenerator(TestCase):
+
+    def test_returns_initial_status(self):
+        initial_status = TicketStatusGenerator.get_initial_status()
+        self.assertIsInstance(initial_status, TicketStatus)
+        self.assertEquals(
+            defaults.SUPPORT_INITIAL_STATUS,
+            initial_status.name
+        )
+        self.assertEquals(
+            defaults.SUPPORT_INITIAL_STATUS_SLUG,
+            initial_status.slug
+        )
+
+    def test_returns_resolved_status(self):
+        resolved_status = TicketStatusGenerator.get_resolved_status()
+        self.assertIsInstance(resolved_status, TicketStatus)
+        self.assertEquals(
+            defaults.SUPPORT_RESOLVED_STATUS,
+            resolved_status.name
+        )
+        self.assertEquals(
+            defaults.SUPPORT_RESOLVED_STATUS_SLUG,
+            resolved_status.slug
+        )
